Fix renaming of registers and #f-ness of classical circuits
This fixes several intertwined issues. First, correctly renames the
register pairing to prevent disassociation of the register names when
renaming internal wires. Second, it fixes renaming when there are two
names that differ only by a number at the end. Previously they could
get renamed to the same name. Finally it changes the checking of output
equality of classical circuits when there is an output wire that is only
present in one circuit. Previously the value #f was used for the
positive and negative wire, which was equivalent to treating this value
as bottom, rather than false.