Add exposed constructivity checking and fix handling of contraints
Previous Contraints could evaluated to `#f` (causing the model to always
pass) if they mentioned variable free in one of the circuit.

The fix is to only check constraints on the first circuit. This has
the added advantage of ensuring that if the constraints connect
the input to the output, then no context are included because
the constraints hold from one circuit but not the other.